requirements.txt
#dj-static==0.0.6 --> Cling
django-heroku
static==1.1.1
Django==1.11.28
django-cors==0.1
django-cors-headers==1.1.0
djangorestframework==3.9.1
gunicorn==19.5.0
wheel==0.24.0
google
Scrapy